refactor: extract fetchRetry into shared utility, fix copyright years
Address PR review feedback:
- Extract duplicated fetchRetry logic from coralogix.js and clickhouse.js
  into a shared createFetchRetry() factory in utils.js
- Fix copyright year to 2026 in new files (constants.js, clickhouse.js,
  clickhouse.test.js)
