+++ src/coralogix.js
@@ -13,10 +13,9 @@
 /* eslint-disable no-await-in-loop */
 
 import path from 'path';
-import wrapFetch from 'fetch-retry';
-import { FetchError, Request } from '@adobe/fetch';
+import { Request } from '@adobe/fetch';
 import { extractFields } from './extract-fields.js';
-import { fetchContext } from './utils.js';
+import { createFetchRetry } from './utils.js';
 import { LOG_LEVEL_MAPPING } from './constants.js';
 
 /**
@@ -38,35 +37,7 @@ import { LOG_LEVEL_MAPPING } from './constants.js';
  * @property {string?} computerName computer name
  */
 
-const { fetch } = fetchContext;
-
-const MOCHA_ENV = (process.env.HELIX_FETCH_FORCE_HTTP1 === 'true');
-
-/**
- * Wrapped fetch that retries on certain conditions.
- */
-const fetchRetry = wrapFetch(fetch, {
-  retryDelay: (attempt) => {
-    if (MOCHA_ENV) {
-      return 1;
-    }
-    /* c8 ignore next */
-    return (2 ** attempt * 1000); // 1000, 2000, 4000
-  },
-  retryOn: async (attempt, error, response) => {
-    const retries = MOCHA_ENV ? 1 /* c8 ignore next */ : 2;
-    if (error) {
-      if (error instanceof FetchError) {
-        return attempt < retries;
-      }
-      throw error;
-    }
-    if (!response.ok) {
-      throw new Error(`Failed to send logs with status ${response.status}: ${await response.text()}`);
-    }
-    return false;
-  },
-});
+const fetchRetry = createFetchRetry('Coralogix');
 
 /**
  * Coralogix logger.
