+++ src/utils.js
@@ -9,7 +9,8 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-import { context, ALPN_HTTP1_1 } from '@adobe/fetch';
+import { context, ALPN_HTTP1_1, FetchError } from '@adobe/fetch';
+import wrapFetch from 'fetch-retry';
 
 /**
  * Our global fetch context that limits the number of sockets used.
@@ -20,6 +21,39 @@ export const fetchContext = context({
   h1: { maxSockets: 512, maxTotalSockets: 768, keepAlive: true },
 });
 
-const { reset } = fetchContext;
+const { reset, fetch } = fetchContext;
 
 export const resetConnections = async () => reset();
+
+const MOCHA_ENV = (process.env.HELIX_FETCH_FORCE_HTTP1 === 'true');
+
+/**
+ * Create a fetch wrapper that retries on FetchError and throws on bad status.
+ *
+ * @param {string} label label used in error messages (e.g. "Coralogix", "ClickHouse")
+ * @returns wrapped fetch function
+ */
+export function createFetchRetry(label) {
+  return wrapFetch(fetch, {
+    retryDelay: (attempt) => {
+      if (MOCHA_ENV) {
+        return 1;
+      }
+      /* c8 ignore next */
+      return (2 ** attempt * 1000); // 1000, 2000, 4000
+    },
+    retryOn: async (attempt, error, response) => {
+      const retries = MOCHA_ENV ? 1 /* c8 ignore next */ : 2;
+      if (error) {
+        if (error instanceof FetchError) {
+          return attempt < retries;
+        }
+        throw error;
+      }
+      if (!response.ok) {
+        throw new Error(`Failed to send logs to ${label} with status ${response.status}: ${await response.text()}`);
+      }
+      return false;
+    },
+  });
+}
